services/transcription_gipformer/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.6
numpy>=1.20.0
soundfile>=0.12.0
sherpa-onnx>=1.10.0
huggingface_hub>=0.20.0
pyannote.audio>=3.3.0
torch>=2.0.0
torchaudio>=2.0.0
